requirements.txt
torchvision==0.2.2.post3
tqdm==4.32.2
torch==1.1.0
matplotlib==3.1.0
numpy==1.16.3
visdom==0.1.8.8
Pillow==6.2.1
